+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changes in mirt 1.21
 
+- penalized term added to EM algorithm estimation subroutines to help keep the covariance matrix
+  of the latent trait parameters positive definite (helps convergence properties of the optimizers, 
+  especially 'L-BFGS-B'). To turn this penalized term off use `technical = list(keep_vcov_PD = FALSE)`
+
 - added `type = 'itemscore'` to `plot()` generic to plot faceted version of the item
   scoring functions